ICP Working ok with 10 or 20 captures, until drift appears. g2o integrated to combat drift

+++ src/GraphOptimizer.h
@@ -10,7 +10,7 @@ class GraphOptimizer
 {
 public:
   /*!Adds a new vertex to the graph. The provided 4x4 matrix will be considered as the pose of the new added vertex. It returns the index of the added vertex.*/
-  virtual int addVertex(Eigen::Matrix4f& pose)=0;
+  virtual int addVertex(Eigen::Matrix4d& pose)=0;
   /*!Adds an edge that defines a spatial constraint between the vertices "fromIdx" and "toIdx" with information matrix that determines the weight of the added edge.*/
   virtual void addEdge(const int fromIdx,const int toIdx,Eigen::Matrix4f& relPose,Eigen::Matrix<double,6,6>& infMatrix)=0;
   /*!Calls the graph optimization process to determine the pose configuration that best satisfies the constraints defined by the edges.*/
